Updated contracts
